requirements.txt
langchain==0.1.0
langchain_community==0.0.10
langsmith==0.0.78
python-dotenv==1.0.0
streamlit==1.28.0
streamlit_feedback==0.1.3
